+++ skbio/util/__init__.py
@@ -8,6 +8,19 @@ scikit-bio, as well as various utility functionality, including I/O and
 unit-testing convenience functions.
 
 
+User configuration utilities
+----------------------------
+
+Configuration options for functional input and output.
+
+.. autosummary::
+   :toctree: generated/
+
+   config
+   get_config
+   set_config
+
+
 Testing utilities
 -----------------
 
@@ -18,6 +31,7 @@ Common functionality to support testing in skbio.
 
    get_data_path
    assert_ordination_results_equal
+   assert_ordination_results_equal_np
    assert_data_frame_almost_equal
 
 
@@ -43,6 +57,15 @@ Decorators
    aliased
    register_aliases
    params_aliased
+
+
+Types
+-----
+
+.. autosummary::
+   :toctree: generated/
+
+   DataTable
 
 
 Miscellaneous utilities
@@ -72,6 +95,7 @@ from ._misc import cardinal_to_ordinal, find_duplicates, safe_md5, get_rng
 from ._testing import (
     get_data_path,
     assert_ordination_results_equal,
+    assert_ordination_results_equal_np,
     assert_data_frame_almost_equal,
     pytestrunner,
 )
@@ -85,6 +109,8 @@ from ._decorator import (
     params_aliased,
 )
 from ._plotting import PlottableMixin
+from .config._config import get_config, set_config
+from ._types import DataTable
 
 __all__ = [
     "cardinal_to_ordinal",
@@ -93,6 +119,7 @@ __all__ = [
     "get_rng",
     "get_data_path",
     "assert_ordination_results_equal",
+    "assert_ordination_results_equal_np",
     "assert_data_frame_almost_equal",
     "pytestrunner",
     "PlottableMixin",
@@ -103,4 +130,7 @@ __all__ = [
     "aliased",
     "register_aliases",
     "params_aliased",
+    "get_config",
+    "set_config",
+    "DataTable",
 ]